runtime: zero-extend Windows error code in syscall_syscalln
asmstdcall now returns the 32-bit Windows error code in the return
register.  cgocall captures this as a signed int32. On 64-bit systems,
casting this directly to uintptr results in sign extension for error
codes with the high bit set (e.g., 0x80092004 becomes
0xffffffff80092004), breaking equality checks against zero-extended
constants.

This CL ensures the error code is zero-extended by casting to uint32
before converting to uintptr.

Fixes #78324

Reviewed-on: https://go-review.googlesource.com/c/go/+/758800
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>

+++ src/runtime/syscall_windows.go
@@ -429,5 +429,5 @@ func syscall_syscalln(fn, n uintptr, args ...uintptr) (r1, r2, err uintptr) {
 	// but it copies the return values into the new M's
 	// so we can read them from there.
 	c = &getg().m.winsyscall
-	return c.R1, c.R2, uintptr(errno)
+	return c.R1, c.R2, uintptr(uint32(errno))
 }
